Improve: Make token verification optional
In order to support old servers with new clients, we need to be able to make the
ID token verification optional. In particular, old services cannot provide a
.well-known URI (which is used to obtain the jwks_uri endpoint).
Hence if there is no wellKnownUri node in the client cfg.xml we only warn, but
do not throw an exception. We then pass keys == null to
IDTokenUtil.verifyAndReadIDToken().
Ideally the whole .well-known URI should only be read once we know that the
token is actually signed. That would automatically provide the backwards
compatibility.

Additionally, logging and warnings for either unsigned tokens or when missing
wellKnown URI (in which case they also cannot be verified) are clarified.