Improve: handle servlet-api centrally
Move from
    javax.servlet:servlet-api:2.5
to
    javax.servlet:javax.servlet-api:3.1.0
Manage the actual version centrally via dependencyManagement so that we don't
need to specify the version for each module separately.
NOTE: in principle we don't use any javax.servlet-api specific things, so we
should not break Tomcat6. If we unexpectedly do, we should move back to
servlet-api and 2.5