Bugfix: idToken should be member of map
The idToken was added to the the RTResponse as the parameters field itself,
instead of as a member of the parameters field. Strictly speaking this would
still have been ok as there aren't any other parameters (yet). The access token
and refresh token are separate fields.
With this fix, the ATResponse2 and RTResponse behave the same.