Bugfix: after rm the list needs to be sorted
When running ls, the list is sorted according to date or id.
Secondly, after removing a client, the list of clients is cleared. Then when
directly afterwards a second client is removed, the list is reloaded, but not
sorted, resulting in an undefined sort order. We fix this by sorting according
to the last used sort order.