Bugfix: table columns need options
Some table columns such as TIMECOLUMN need to be created with special options
such as DEFAULT CURRENT_TIMESTAMP. In order to allow for this we add an options
field which is inserted after the table type.
Additionally, instead of logging only the name and type of the added column, we
log the full statement.