Improve: use UTF-8 for POST
Assume posted data is UTF-8 encoded, which is more likely than other encodings.
Note that this has to be done before the request parameters are read, hence at
the beginning of doPost().
We (currently) do this in the AbstractServlet which is parent for all servlets
including registration ones for which this is the most important.